Validates org existence when creating a pet
Ensures that an organization exists before allowing a pet to be created and associated with it.

This change adds validation to the `CreatePetUseCase` to check if the provided `orgId` corresponds to an actual organization in the system. If the organization doesn't exist, an error is thrown, preventing the creation of the pet.

Additionally, this commit:
- Defines abstract method `findByID` in `ORGSRepo`
- Implements `findByID` in `InMemoryOrgsRepo` and `PrismaOrgsRepo`
- Registers the pet routes in the `server.ts` file.
- Renames the org routes to pet routes in `routes.ts`